+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/implementation/jms/ServiceBusJmsConnectionFactoryConfiguration.java
@@ -3,10 +3,17 @@
 
 package com.azure.spring.cloud.autoconfigure.implementation.jms;
 
+import com.azure.core.credential.TokenCredential;
+import com.azure.identity.extensions.implementation.credential.TokenCredentialProviderOptions;
+import com.azure.identity.extensions.implementation.credential.provider.TokenCredentialProvider;
+import com.azure.identity.extensions.implementation.enums.AuthProperty;
 import com.azure.servicebus.jms.ServiceBusJmsConnectionFactory;
 import com.azure.spring.cloud.autoconfigure.implementation.jms.properties.AzureServiceBusJmsProperties;
 import com.azure.spring.cloud.autoconfigure.jms.AzureServiceBusJmsConnectionFactoryCustomizer;
+import com.azure.spring.cloud.service.implementation.identity.credential.provider.SpringTokenCredentialProvider;
 import jakarta.jms.ConnectionFactory;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.apache.commons.pool2.PooledObject;
 import org.messaginghub.pooled.jms.JmsPoolConnectionFactory;
 import org.springframework.beans.factory.ObjectProvider;
@@ -17,28 +24,70 @@ import org.springframework.boot.autoconfigure.jms.JmsPoolConnectionFactoryFactor
 import org.springframework.boot.autoconfigure.jms.JmsProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.DependsOn;
+import org.springframework.context.support.GenericApplicationContext;
 import org.springframework.jms.connection.CachingConnectionFactory;
+import org.springframework.util.StringUtils;
 
+import java.util.Properties;
 import java.util.stream.Collectors;
+
+import static com.azure.spring.cloud.service.implementation.identity.credential.provider.SpringTokenCredentialProvider.PASSWORDLESS_TOKEN_CREDENTIAL_BEAN_NAME;
 
 @Configuration(proxyBeanMethods = false)
 @ConditionalOnMissingBean(ConnectionFactory.class)
-class ServiceBusJmsConnectionFactoryConfiguration {
+class ServiceBusJmsConnectionFactoryConfiguration  {
 
-    private static ServiceBusJmsConnectionFactory createJmsConnectionFactory(AzureServiceBusJmsProperties properties,
+    private static final Log LOGGER = LogFactory.getLog(ServiceBusJmsConnectionFactoryConfiguration.class);
+    private final GenericApplicationContext applicationContext;
+
+    ServiceBusJmsConnectionFactoryConfiguration(GenericApplicationContext applicationContext) {
+        this.applicationContext = applicationContext;
+    }
+
+    private ServiceBusJmsConnectionFactory createJmsConnectionFactory(AzureServiceBusJmsProperties jmsProperties,
                                                                              ObjectProvider<AzureServiceBusJmsConnectionFactoryCustomizer> factoryCustomizers) {
-        return new ServiceBusJmsConnectionFactoryFactory(properties,
+        TokenCredentialProvider tokenCredentialProvider = getPasswordlessTokenCredentialProvider(jmsProperties);
+        return new ServiceBusJmsConnectionFactoryFactory(tokenCredentialProvider, jmsProperties,
             factoryCustomizers.orderedStream().collect(Collectors.toList()))
             .createConnectionFactory(ServiceBusJmsConnectionFactory.class);
+    }
+
+    private TokenCredentialProvider getPasswordlessTokenCredentialProvider(AzureServiceBusJmsProperties serviceBusJmsProperties) {
+        if (!serviceBusJmsProperties.isPasswordlessEnabled()) {
+            LOGGER.debug("Feature passwordless authentication is not enabled(" + AzureServiceBusJmsProperties.PREFIX + ".passwordless-enabled=false), "
+                + "skip enhancing Service Bus JMS properties.");
+            return null;
+        }
+
+        Properties properties = serviceBusJmsProperties.toPasswordlessProperties();
+        String tokenCredentialBeanName = serviceBusJmsProperties.getCredential().getTokenCredentialBeanName();
+        if (StringUtils.hasText(tokenCredentialBeanName)) {
+            AuthProperty.TOKEN_CREDENTIAL_BEAN_NAME.setProperty(properties, tokenCredentialBeanName);
+        } else {
+            TokenCredentialProvider tokenCredentialProvider = TokenCredentialProvider.createDefault(new TokenCredentialProviderOptions(properties));
+            TokenCredential tokenCredential = tokenCredentialProvider.get();
+
+            tokenCredentialBeanName = PASSWORDLESS_TOKEN_CREDENTIAL_BEAN_NAME + "." + AzureServiceBusJmsProperties.PREFIX;
+            AuthProperty.TOKEN_CREDENTIAL_BEAN_NAME.setProperty(properties, tokenCredentialBeanName);
+            applicationContext.registerBean(tokenCredentialBeanName, TokenCredential.class, () -> tokenCredential);
+        }
+
+        AuthProperty.TOKEN_CREDENTIAL_PROVIDER_CLASS_NAME.setProperty(properties, SpringTokenCredentialProvider.class.getName());
+        AuthProperty.AUTHORITY_HOST.setProperty(properties, serviceBusJmsProperties.getProfile().getEnvironment().getActiveDirectoryEndpoint());
+
+        return TokenCredentialProvider.createDefault(new TokenCredentialProviderOptions(properties));
     }
 
     @Configuration(proxyBeanMethods = false)
     @ConditionalOnProperty(prefix = "spring.jms.servicebus.pool", name = "enabled", havingValue = "false",
         matchIfMissing = true)
-    static class SimpleConnectionFactoryConfiguration {
+    class SimpleConnectionFactoryConfiguration {
+
 
         @Bean
         @ConditionalOnProperty(prefix = "spring.jms.cache", name = "enabled", havingValue = "false")
+        @DependsOn("springTokenCredentialProviderContextProvider")
         ServiceBusJmsConnectionFactory jmsConnectionFactory(AzureServiceBusJmsProperties properties,
                                                             ObjectProvider<AzureServiceBusJmsConnectionFactoryCustomizer> factoryCustomizers) {
             return createJmsConnectionFactory(properties, factoryCustomizers);
@@ -48,9 +97,10 @@ class ServiceBusJmsConnectionFactoryConfiguration {
         @ConditionalOnClass(CachingConnectionFactory.class)
         @ConditionalOnProperty(prefix = "spring.jms.cache", name = "enabled", havingValue = "true",
             matchIfMissing = true)
-        static class CachingConnectionFactoryConfiguration {
+        class CachingConnectionFactoryConfiguration {
 
             @Bean
+            @DependsOn("springTokenCredentialProviderContextProvider")
             CachingConnectionFactory jmsConnectionFactory(JmsProperties jmsProperties,
                                                           AzureServiceBusJmsProperties properties,
                                                           ObjectProvider<AzureServiceBusJmsConnectionFactoryCustomizer> factoryCustomizers) {
@@ -68,10 +118,11 @@ class ServiceBusJmsConnectionFactoryConfiguration {
 
     @Configuration(proxyBeanMethods = false)
     @ConditionalOnClass({ JmsPoolConnectionFactory.class, PooledObject.class })
-    static class PooledConnectionFactoryConfiguration {
+    class PooledConnectionFactoryConfiguration {
 
         @Bean(destroyMethod = "stop")
         @ConditionalOnProperty(prefix = "spring.jms.servicebus.pool", name = "enabled", havingValue = "true")
+        @DependsOn("springTokenCredentialProviderContextProvider")
         JmsPoolConnectionFactory jmsPoolConnectionFactory(AzureServiceBusJmsProperties properties,
                                                           ObjectProvider<AzureServiceBusJmsConnectionFactoryCustomizer> factoryCustomizers) {
             ServiceBusJmsConnectionFactory factory = createJmsConnectionFactory(properties, factoryCustomizers);