+++ sdk/spring/spring-cloud-azure-core/src/main/java/com/azure/spring/cloud/core/provider/authentication/TokenCredentialOptionsProvider.java
@@ -62,6 +62,11 @@ public interface TokenCredentialOptionsProvider {
          */
         boolean isManagedIdentityEnabled();
 
+        /**
+         * Get the custom {@link com.azure.core.credential.TokenCredential} bean name.
+         * @return the token credential bean name.
+         */
+        String getTokenCredentialBeanName();
     }
 
 }